Re structure the api
