+++ run.py
@@ -1,6 +1,4 @@
-from app.routes import app
+from app import app
 
 if __name__ == "__main__":
     app.run(debug=True)
-
-
